test added to github actions, moved files
