+++ backend/src/jobs/tweet/schedule-tweets-model.ts
@@ -1,4 +1,4 @@
-import prisma from '../../../../lib/prisma';
+import prisma from '../../../lib/prisma';
 import { ScheduledTweets } from './schedule-tweets';
 
 /**
