+++ backend/src/jobs/tweet/schedule-tweets.d.ts
@@ -1,5 +1,5 @@
-import { Tweets } from '../../../routes/tweets/tweets';
-import { TwitterOAuth } from '../../../routes/oauths/oauths';
+import { Tweets } from '../../routes/tweets/tweets';
+import { TwitterOAuth } from '../../routes/oauths/oauths';
 
 interface ScheduledTweet {
 	id: bigint; // twitter user id
